+++ src/components/general-widgets/Review/component.spec.js
@@ -14,7 +14,6 @@ import { GridRow } from 'layout/GridRow';
 import { GridColumn } from 'layout/GridColumn';
 import { Quote } from 'elements/Quote';
 import { Subheading } from 'typography/Subheading';
-import { VerticalGutters } from 'layout/VerticalGutters';
 
 import { getReviewerNameAndLocationString } from './utils/getReviewerNameAndLocationString';
 import { getReviewerCategoryAndStayDateString } from './utils/getReviewerCategoryAndStayDateString';
@@ -40,17 +39,10 @@ const getReview = additionalProps =>
   shallow(<Review {...requiredProps} {...additionalProps} />);
 
 describe('<Review />', () => {
-  it('should have `VerticalGutters` component as a wrapper', () => {
+  it('should have `Card` component as a wrapper', () => {
     const wrapper = getReview();
 
-    expectComponentToBe(wrapper, VerticalGutters);
-  });
-  describe('the `VerticalGutters` component', () => {
-    it('should have `Card` as its only children', () => {
-      const wrapper = getReview();
-
-      expectComponentToHaveChildren(wrapper, Card);
-    });
+    expectComponentToBe(wrapper, Card);
   });
 
   describe('the first `Card` component', () => {