+++ src/index.js
@@ -2,6 +2,8 @@
 export { Footer } from './components/collections/Footer';
 export { Form } from './components/collections/Form';
 export { Header } from './components/collections/Header';
+export { Hero } from './components/collections/Hero';
+export { HomepageHero } from './components/collections/HomepageHero';
 export { InputGroup } from './components/collections/InputGroup';
 
 // Elements
@@ -70,9 +72,6 @@ export { Video } from './components/media/Video';
 export { Amenities } from './components/property-page-widgets/Amenities';
 export { Availability } from './components/property-page-widgets/Availability';
 export { Description } from './components/property-page-widgets/Description';
-export {
-  PropertySearchBar,
-} from './components/property-page-widgets/PropertySearchBar';
 export { HostProfile } from './components/property-page-widgets/HostProfile';
 export { KeyFacts } from './components/property-page-widgets/KeyFacts';
 export { Location } from './components/property-page-widgets/Location';